Rename custom-script into script
