+++ components/script.tsx
